# swagger.yml
basePath: /
consumes:
    - application/json
definitions:
    ACL:
        additionalProperties:
            format: uint8
            type: integer
        description: ACL - the ACL of other nodes in a NetworkACL for a single unique node
        type: object
        x-go-package: github.com/gravitl/netmaker/logic/acls
    ACLContainer:
        additionalProperties:
            $ref: '#/definitions/ACL'
        description: ACLContainer - the total list of all node's ACL in a given network
        type: object
        x-go-package: github.com/gravitl/netmaker/logic/acls
...
    Network:
        description: |-
            Network Struct - contains info for a given unique network
            At  some point, need to replace all instances of Name with something else like  Identifier
        properties:
            addressrange:
                type: string
                x-go-name: AddressRange
            addressrange6:
                type: string
                x-go-name: AddressRange6
            allowmanualsignup:
                type: string
                x-go-name: AllowManualSignUp
            defaultacl:
                type: string
                x-go-name: DefaultACL
            defaultinterface:
                type: string
                x-go-name: DefaultInterface
            defaultkeepalive:
                format: int32
                type: integer
                x-go-name: DefaultKeepalive
            defaultlistenport:
                format: int32
                type: integer
                x-go-name: DefaultListenPort
            defaultmtu:
                format: int32
                type: integer
                x-go-name: DefaultMTU
            defaultpostdown:
                type: string
                x-go-name: DefaultPostDown
            defaultudpholepunch:
                type: string
                x-go-name: DefaultUDPHolePunch
            isipv4:
                type: string
                x-go-name: IsIPv4
            isipv6:
                type: string
                x-go-name: IsIPv6
            netid:
                type: string
                x-go-name: NetID
            networklastmodified:
                format: int64
                type: integer
                x-go-name: NetworkLastModified
            nodelimit:
                format: int32
                type: integer
                x-go-name: NodeLimit
            nodeslastmodified:
                format: int64
                type: integer
                x-go-name: NodesLastModified
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
    SuccessResponse:
        properties:
            Code:
                format: int64
                type: integer
            Message:
                type: string
            Response: {}
        title: SuccessResponse is struct for sending error message with code.
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
host: api.demo.netmaker.io
info:
    description: |-
        # API Usage

        Most actions that can be performed via API can be performed via UI. We recommend managing your networks using the official netmaker-ui project. However, Netmaker can also be run without the UI, and all functions can be achieved via API calls. If your use case requires using Netmaker without the UI or you need to do some troubleshooting/advanced configuration, using the API directly may help.

        # Authentication

        API calls must be authenticated via a header of the format -H “Authorization: Bearer <YOUR_SECRET_KEY>” There are two methods to obtain YOUR_SECRET_KEY: 1. Using the masterkey. By default, this value is “secret key,” but you should change this on your instance and keep it secure. This value can be set via env var at startup or in a config file (config/environments/< env >.yaml). See the [Netmaker](https://docs.netmaker.org/index.html) documentation for more details. 2. Using a JWT received for a node. This can be retrieved by calling the /api/nodes/<network>/authenticate endpoint, as documented below.
    title: Netmaker
    version: 0.21.1
paths:
...
    /api/networks:
        get:
            operationId: getNetworks
            responses:
                "200":
                    $ref: '#/responses/getNetworksSliceResponse'
            schemes:
                - https
            summary: Lists all networks.
            tags:
                - networks
        post:
            operationId: createNetwork
            parameters:
                - description: Network
                  in: body
                  name: network
                  schema:
                    $ref: '#/definitions/Network'
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Create a network.
            tags:
                - networks
        put:
            operationId: updateNetwork
            parameters:
                - description: Network
                  in: body
                  name: network
                  schema:
                    $ref: '#/definitions/Network'
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Update pro settings for a network.
            tags:
                - networks
    /api/networks/{networkname}:
        delete:
            operationId: deleteNetwork
            parameters:
                - description: 'name: network name'
                  in: path
                  name: networkname
                  required: true
                  type: string
                  x-go-name: Networkname
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            schemes:
                - https
            summary: Delete a network.  Will not delete if there are any nodes that belong to the network.
            tags:
                - networks
        get:
            operationId: getNetwork
            parameters:
                - description: 'name: network name'
                  in: path
                  name: networkname
                  required: true
                  type: string
                  x-go-name: Networkname
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Get a network.
            tags:
                - networks
    /api/networks/{networkname}/acls:
        get:
            operationId: getNetworkACL
            parameters:
                - description: 'name: network name'
                  in: path
                  name: networkname
                  required: true
                  type: string
                  x-go-name: Networkname
                - description: ACL Container
                  in: body
                  name: acl_container
                  schema:
                    $ref: '#/definitions/ACLContainer'
                  x-go-name: ACLContainer
            responses:
                "200":
                    $ref: '#/responses/aclContainerResponse'
            schemes:
                - https
            summary: Get a network ACL (Access Control List).
            tags:
                - networks
        put:
            operationId: updateNetworkACL
            parameters:
                - description: 'name: network name'
                  in: path
                  name: networkname
                  required: true
                  type: string
                  x-go-name: Networkname
                - description: ACL Container
                  in: body
                  name: acl_container
                  schema:
                    $ref: '#/definitions/ACLContainer'
                  x-go-name: ACLContainer
            responses:
                "200":
                    $ref: '#/responses/aclContainerResponse'
            schemes:
                - https
            summary: Update a network ACL (Access Control List).
            tags:
                - networks
...
produces:
    - application/json
responses:
...
    aclContainerResponse:
        description: ""
        schema:
            $ref: '#/definitions/ACLContainer'
...
    getNetworksSliceResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/Network'
            type: array
...
    networkBodyResponse:
        description: ""
        schema:
            $ref: '#/definitions/Network'
...
    successResponse:
        description: ""
        schema:
            $ref: '#/definitions/SuccessResponse'
...
schemes:
    - https
swagger: "2.0"